third_party/WebKit/LayoutTests/http/tests/serviceworker/unregister-controller.html

<!DOCTYPE html>
<script src="../resources/testharness.js"></script>
<script src="../resources/testharnessreport.js"></script>
<script src="resources/test-helpers.js"></script>
<script>
var worker_url = 'resources/simple-intercept-worker.js';

async_test(function(t) {
  var scope =
      'resources/unregister-controller-page.html?load-before-unregister';
  var worker;
  var frame_window;
  var controller;

  service_worker_unregister_and_register(t, worker_url, scope)
    .then(function(registered_worker) {
        worker = registered_worker;
        return wait_for_state(t, worker, 'activated');
      })
    .then(function() {
        return with_iframe(scope);
      })
    .then(function(frame) {
        frame_window = frame.contentWindow;
        controller = frame_window.navigator.serviceWorker.controller;
        assert_true(controller instanceof frame_window.ServiceWorker,
                    'document should load with a controller');
        return navigator.serviceWorker.unregister(scope);
      })
    .then(function() {
        assert_equals(frame_window.navigator.serviceWorker.controller,
                      controller,
                      'unregistration should not modify controller');
        return frame_window.fetch_url('simple.txt');
      })
    .then(function(response) {
        assert_equals(response, 'intercepted by service worker',
                      'controller should intercept requests');
        t.done();
      })
    .catch(unreached_rejection(t));
}, 'Unregister does not affect existing controller');

async_test(function(t) {
  var scope =
      'resources/unregister-controller-page.html?load-after-unregister';

  service_worker_unregister_and_register(t, worker_url, scope)
    .then(function(worker) {
        return wait_for_state(t, worker, 'activated');
      })
    .then(function() {
        return navigator.serviceWorker.unregister(scope);
      })
    .then(function() {
        return with_iframe(scope);
      })
    .then(function(frame) {
        var frame_window = frame.contentWindow;
        assert_equals(frame_window.navigator.serviceWorker.controller, null,
                      'document should not have a controller');
        return frame_window.fetch_url('simple.txt');
      })
    .then(function(response) {
        assert_equals(response, 'a simple text file\n',
                      'requests should not be intercepted');
        t.done();
      })
    .catch(unreached_rejection(t));
}, 'Unregister prevents control of subsequent navigations');

async_test(function(t) {
    var scope =
        'scope/no-new-controllee-even-if-registration-is-still-used';

    service_worker_unregister_and_register(t, worker_url, scope)
      .then(function(registered_worker) {
          return wait_for_state(t, registered_worker, 'activated');
        })
      .then(function() {
          return with_iframe(scope);
        })
      .then(function(frame) {
          return navigator.serviceWorker.unregister(scope);
        })
      .then(function() {
          return with_iframe(scope);
        })
      .then(function(frame) {
          assert_equals(frame.contentWindow.navigator.serviceWorker.controller,
                        null,
                        'document should not have a controller');
          t.done();
        })
      .catch(unreached_rejection(t));
  }, 'Unregister prevents new controllee even if registration is still in use');
</script>
